+++ frontend/src/main.ts
@@ -12,6 +12,7 @@ import './assets/css/theme.css'
 import './assets/css/main.css'
 import './assets/css/dashboard.css'
 import './assets/css/layout.css'
+import './assets/css/module.css'
 
 /* Basic CSS for apps built with Ionic */
 import '@ionic/vue/css/normalize.css';